Refactors helper testing
