+++ app/helper/prettyPrice.js
@@ -1,6 +1,9 @@
-var prettyNumber = require("./prettyNumber");
+const prettyNumber = require("helper/prettyNumber");
+const ensure = require("helper/ensure");
 
-function prettyPrice(cents) {
+module.exports = function prettyPrice(cents) {
+  ensure(cents, "number");
+
   var price = (cents / 100).toFixed(2);
 
   if (price.slice(-2) === "00") price = price.slice(0, -3);
@@ -8,15 +11,4 @@ function prettyPrice(cents) {
   price = prettyNumber(price);
 
   return "$" + price;
-}
-
-var is = require("./_is")(prettyPrice);
-
-is(2134, "$21.34");
-is(2000, "$20");
-is(200, "$2");
-is(210, "$2.10");
-is(03, "$0.03");
-is(30, "$0.30");
-
-module.exports = prettyPrice;
+};
