+++ app/helper/tests/urlNormalizer.js
@@ -1,0 +1,14 @@
+describe("urlNormalizer", function () {
+  const urlNormalizer = require("helper/urlNormalizer");
+  const is = require("./util/is")(urlNormalizer);
+
+  it("works", function () {
+    is("", "");
+    is("/", "/");
+    is("//", "/");
+    is("/fo/", "/fo");
+    is("http://blot.im/foo/bar", "/foo/bar");
+    is("/foo/bar/", "/foo/bar");
+    is("foo/bar/", "/foo/bar");
+  });
+});
