+++ app/helper/withoutExtension.js
@@ -1,8 +1,8 @@
-var ensure = require("./ensure");
-var makeSlug = require("./makeSlug");
+var ensure = require("helper/ensure");
+var makeSlug = require("helper/makeSlug");
 
-// Takes a path, return the path with the file extension
-function withoutExtension(path) {
+module.exports = function withoutExtension(path) {
+  // Takes a path, return the path with the file extension
   ensure(path, "string");
 
   if (!path) return "";
@@ -14,7 +14,7 @@ function withoutExtension(path) {
   path = names.join("/");
 
   return path;
-}
+};
 
 function strip(name) {
   if (name.indexOf(".") > -1) {
@@ -26,14 +26,3 @@ function strip(name) {
 
   return name;
 }
-
-var is = require("./_is")(withoutExtension);
-
-is("", "");
-is("bar.txt", "bar");
-is("foo", "foo");
-is("Foo.bar.baz", "Foo.bar");
-is("/foo/bar/baz.txt", "/foo/bar/baz");
-is("/foo/bar.foo.txt", "/foo/bar.foo");
-is("foo.bar/bar.baz", "foo.bar/bar");
-module.exports = withoutExtension;
